Start to reduce/refactor arithmetic ...
Remove stuff from mathics.builtin.arithmetic that does not belong there:

* _MPMathFunction -> MPMathFunction and move to mathics.buitin.base
* _MPMathMultiFunction -> MPMathFunction and move to mathics.builtin.base
* Abs, Piecewise, RealAbs, RealSign, Sign moved to numeric to follow WMA
  organization better

The corresponding eval routines will be gone over in another PR. This
one is already large.

Url's gone over to make this not exceed standard line limit.
Note that the formatting has been gone over to follow the existing
pattern that we have been using.

cythonization in mathics.builtin class files removed. It is not
clear this has benefit in modern Pythons, especially in this kind of builtin function
and all of this needs to be retested if not rethought.
