+++ mathics/builtin/numeric.py
@@ -1,4 +1,3 @@
-# cython: language_level=3
 # -*- coding: utf-8 -*-
 
 # Note: docstring is flowed in documentation. Line breaks in the
@@ -12,16 +11,47 @@ Support for approximate real numbers and exact real numbers represented \
 in algebraic or symbolic form.
 """
 
+from typing import Optional
+
 import sympy
 
-from mathics.builtin.base import Builtin
-from mathics.core.atoms import Complex, Integer, Integer0, Rational, Real
-from mathics.core.attributes import A_LISTABLE, A_NUMERIC_FUNCTION, A_PROTECTED
+from mathics.builtin.base import Builtin, MPMathFunction, SympyFunction
+from mathics.builtin.inference import evaluate_predicate
+from mathics.core.atoms import (
+    Complex,
+    Integer,
+    Integer0,
+    IntegerM1,
+    Number,
+    Rational,
+    Real,
+)
+from mathics.core.attributes import (
+    A_HOLD_ALL,
+    A_LISTABLE,
+    A_NUMERIC_FUNCTION,
+    A_PROTECTED,
+)
 from mathics.core.convert.sympy import from_sympy
+from mathics.core.element import BaseElement
 from mathics.core.evaluation import Evaluation
 from mathics.core.expression import Expression
 from mathics.core.number import MACHINE_EPSILON
-from mathics.core.symbols import SymbolDivide, SymbolMachinePrecision, SymbolTimes
+from mathics.core.symbols import (
+    Symbol,
+    SymbolDivide,
+    SymbolFalse,
+    SymbolMachinePrecision,
+    SymbolTimes,
+    SymbolTrue,
+)
+from mathics.core.systemsymbols import SymbolPiecewise
+from mathics.eval.arithmetic import (
+    eval_Abs,
+    eval_negate_number,
+    eval_RealSign,
+    eval_Sign,
+)
 from mathics.eval.nevaluator import eval_NValues
 
 
@@ -43,6 +73,54 @@ def chop(expr, delta=10.0 ** (-10.0)):
             chop(expr.head), *[chop(element) for element in expr.elements]
         )
     return expr
+
+
+class Abs(MPMathFunction):
+    """
+    <url>
+    :Absolute value:
+    https://en.wikipedia.org/wiki/Absolute_value</url> (<url>
+    :SymPy:
+    https://docs.sympy.org/latest/modules/functions
+    /elementary.html#sympy.functions.elementary.complexes.Abs</url>, <url>
+    :WMA: https://reference.wolfram.com/language/ref/Abs</url>)
+
+    <dl>
+      <dt>'Abs[$x$]'
+      <dd>returns the absolute value of $x$.
+    </dl>
+
+    >> Abs[-3]
+     = 3
+
+    >> Plot[Abs[x], {x, -4, 4}]
+     = -Graphics-
+
+    'Abs' returns the magnitude of complex numbers:
+    >> Abs[3 + I]
+     = Sqrt[10]
+    >> Abs[3.0 + I]
+     = 3.16228
+
+    All of the below evaluate to Infinity:
+
+    >> Abs[Infinity] == Abs[I Infinity] == Abs[ComplexInfinity]
+     = True
+    """
+
+    mpmath_name = "fabs"  # mpmath actually uses python abs(x) / x.__abs__()
+    rules = {
+        "Abs[Undefined]": "Undefined",
+    }
+    summary_text = "absolute value of a number"
+    sympy_name = "Abs"
+
+    def eval(self, x, evaluation: Evaluation):
+        "Abs[x_]"
+        result = eval_Abs(x)
+        if result is not None:
+            return result
+        return super(Abs, self).eval(x, evaluation)
 
 
 class Chop(Builtin):
@@ -252,6 +330,101 @@ class N(Builtin):
         return eval_NValues(expr, SymbolMachinePrecision, evaluation)
 
 
+class Piecewise(SympyFunction):
+    """
+    <url>:WMA link:https://reference.wolfram.com/language/ref/Piecewise.html</url>
+
+    <dl>
+      <dt>'Piecewise[{{expr1, cond1}, ...}]'
+      <dd>represents a piecewise function.
+
+      <dt>'Piecewise[{{expr1, cond1}, ...}, expr]'
+      <dd>represents a piecewise function with default 'expr'.
+    </dl>
+
+    Heaviside function
+    >> Piecewise[{{0, x <= 0}}, 1]
+     = Piecewise[{{0, x <= 0}}, 1]
+
+    ## D[%, x]
+    ## Piecewise({{0, Or[x < 0, x > 0]}}, Indeterminate).
+
+    >> Integrate[Piecewise[{{1, x <= 0}, {-1, x > 0}}], x]
+     = Piecewise[{{x, x <= 0}}, -x]
+
+    >> Integrate[Piecewise[{{1, x <= 0}, {-1, x > 0}}], {x, -1, 2}]
+     = -1
+
+    Piecewise defaults to 0 if no other case is matching.
+    >> Piecewise[{{1, False}}]
+     = 0
+
+    >> Plot[Piecewise[{{Log[x], x > 0}, {x*-0.5, x < 0}}], {x, -1, 1}]
+     = -Graphics-
+
+    >> Piecewise[{{0 ^ 0, False}}, -1]
+     = -1
+    """
+
+    summary_text = "an arbitrary piecewise function"
+    sympy_name = "Piecewise"
+
+    attributes = A_HOLD_ALL | A_PROTECTED
+
+    def eval(self, items, evaluation: Evaluation):
+        "%(name)s[items__]"
+        result = self.to_sympy(
+            Expression(SymbolPiecewise, *items.get_sequence()), evaluation=evaluation
+        )
+        if result is None:
+            return
+        if not isinstance(result, sympy.Piecewise):
+            result = from_sympy(result)
+            return result
+
+    def to_sympy(self, expr, **kwargs):
+        elements = expr.elements
+        evaluation = kwargs.get("evaluation", None)
+        if len(elements) not in (1, 2):
+            return
+
+        sympy_cases = []
+        for case in elements[0].elements:
+            if case.get_head_name() != "System`List":
+                return
+            if len(case.elements) != 2:
+                return
+            then, cond = case.elements
+            if evaluation:
+                cond = evaluate_predicate(cond, evaluation)
+
+            sympy_cond = None
+            if isinstance(cond, Symbol):
+                if cond is SymbolTrue:
+                    sympy_cond = True
+                elif cond is SymbolFalse:
+                    sympy_cond = False
+            if sympy_cond is None:
+                sympy_cond = cond.to_sympy(**kwargs)
+                if not (sympy_cond.is_Relational or sympy_cond.is_Boolean):
+                    return
+
+            sympy_cases.append((then.to_sympy(**kwargs), sympy_cond))
+
+        if len(elements) == 2:  # default case
+            sympy_cases.append((elements[1].to_sympy(**kwargs), True))
+        else:
+            sympy_cases.append((Integer0.to_sympy(**kwargs), True))
+
+        return sympy.Piecewise(*sympy_cases)
+
+    def from_sympy(self, sympy_name, args):
+        # Hack to get around weird sympy.Piecewise 'otherwise' behaviour
+        if str(args[-1].elements[1]).startswith("System`_True__Dummy_"):
+            args[-1].elements[1] = SymbolTrue
+        return Expression(self.get_name(), args)
+
+
 class Rationalize(Builtin):
     """
     <url>:WMA link:
@@ -395,6 +568,86 @@ class Rationalize(Builtin):
         return result
 
 
+class RealAbs(Builtin):
+    """
+    <url>:Abs (Real):
+    https://en.wikipedia.org/wiki/Absolute_value</url> (<url>
+    :WMA link:
+    https://reference.wolfram.com/language/ref/RealAbs.html</url>)
+
+    <dl>
+      <dt>'RealAbs[$x$]'
+      <dd>returns the absolute value of a real number $x$.
+    </dl>
+    'RealAbs' is also known as modulus. It is evaluated if $x$ can be compared \
+    with $0$.
+
+    >> RealAbs[-3.]
+     = 3.
+    'RealAbs[$z$]' is left unevaluated for complex $z$:
+    >> RealAbs[2. + 3. I]
+     = RealAbs[2. + 3. I]
+    >> D[RealAbs[x ^ 2], x]
+     = 2 x ^ 3 / RealAbs[x ^ 2]
+    """
+
+    attributes = A_LISTABLE | A_NUMERIC_FUNCTION | A_PROTECTED
+    rules = {
+        "D[RealAbs[x_],x_]": "x/RealAbs[x]",
+        "Integrate[RealAbs[x_],x_]": "1/2 x RealAbs[x]",
+        "Integrate[RealAbs[u_],{u_,a_,b_}]": "1/2 b RealAbs[b]-1/2 a RealAbs[a]",
+    }
+    summary_text = "real absolute value"
+
+    def eval(self, x: BaseElement, evaluation: Evaluation):
+        """RealAbs[x_]"""
+        real_sign = eval_RealSign(x)
+        if real_sign is IntegerM1:
+            return eval_negate_number(x)
+        if real_sign is None:
+            return
+        return x
+
+
+class RealSign(Builtin):
+    """
+    <url>:Sign function:
+    https://en.wikipedia.org/wiki/Sign_function</url> (<url>
+    :WMA link:
+    https://reference.wolfram.com/language/ref/RealSign.html</url>)
+
+    <dl>
+      <dt>'RealSign[$x$]'
+      <dd>returns -1, 0 or 1 depending on whether $x$ is negative,
+      zero or positive.
+    </dl>
+    'RealSign' is also known as $sgn$ or $signum$ function.
+
+    >> RealSign[-3.]
+     = -1
+    'RealSign[$z$]' is left unevaluated for complex $z$:
+    >> RealSign[2. + 3. I]
+     = RealSign[2. + 3. I]
+
+    >> D[RealSign[x^2],x]
+     = 2 x Piecewise[{{0, x ^ 2 != 0}}, Indeterminate]
+    >> Integrate[RealSign[u],{u,0,x}]
+     = RealAbs[x]
+    """
+
+    attributes = A_LISTABLE | A_NUMERIC_FUNCTION | A_PROTECTED
+    rules = {
+        "D[RealSign[x_],x_]": "Piecewise[{{0, x!=0}}, Indeterminate]",
+        "Integrate[RealSign[x_],x_]": "RealAbs[x]",
+        "Integrate[RealSign[u_],{u_, a_, b_}]": "RealAbs[b]-RealSign[a]",
+    }
+    summary_text = "real sign"
+
+    def eval(self, x: Number, evaluation: Evaluation) -> Optional[Integer]:
+        """RealSign[x_]"""
+        return eval_RealSign(x)
+
+
 class RealValuedNumberQ(Builtin):
     # No docstring since this is internal and it will mess up documentation.
     # FIXME: Perhaps in future we will have a more explicite way to indicate not
@@ -493,3 +746,69 @@ class Round(Builtin):
             n = round(n)
         n = int(n)
         return Expression(SymbolTimes, Integer(n), k)
+
+
+class Sign(SympyFunction):
+    """
+    <url>
+    :Sign:
+    https://en.wikipedia.org/wiki/Sign_function</url> (<url>
+    :WMA link:
+    https://reference.wolfram.com/language/ref/Sign.html</url>)
+
+    <dl>
+      <dt>'Sign[$x$]'
+      <dd>return -1, 0, or 1 depending on whether $x$ is negative, zero, or positive.
+    </dl>
+
+    >> Sign[19]
+     = 1
+    >> Sign[-6]
+     = -1
+    >> Sign[0]
+     = 0
+    >> Sign[{-5, -10, 15, 20, 0}]
+     = {-1, -1, 1, 1, 0}
+    #> Sign[{1, 2.3, 4/5, {-6.7, 0}, {8/9, -10}}]
+     = {1, 1, 1, {-1, 0}, {1, -1}}
+    >> Sign[3 - 4*I]
+     = 3 / 5 - 4 I / 5
+    #> Sign[1 - 4*I] == (1/17 - 4 I/17) Sqrt[17]
+     = True
+    #> Sign[4, 5, 6]
+     : Sign called with 3 arguments; 1 argument is expected.
+     = Sign[4, 5, 6]
+    #> Sign["20"]
+     = Sign[20]
+    """
+
+    summary_text = "complex sign of a number"
+    sympy_name = "sign"
+    # mpmath_name = 'sign'
+
+    attributes = A_LISTABLE | A_NUMERIC_FUNCTION | A_PROTECTED
+
+    messages = {
+        "argx": "Sign called with `1` arguments; 1 argument is expected.",
+    }
+
+    rules = {
+        "Sign[Power[a_, b_]]": "Power[Sign[a], b]",
+    }
+
+    def eval(self, x, evaluation: Evaluation):
+        "Sign[x_]"
+        result = eval_Sign(x)
+        if result is not None:
+            return result
+        # return None
+
+        sympy_x = x.to_sympy()
+        if sympy_x is None:
+            return None
+        # Unhandled cases. Use sympy
+        return super(Sign, self).eval(x, evaluation)
+
+    def eval_error(self, x, seqs, evaluation: Evaluation):
+        "Sign[x_, seqs__]"
+        evaluation.message("Sign", "argx", Integer(len(seqs.get_sequence()) + 1))
